Update config builder
(cherry picked from commit fb86e37c931ca4268dffa3eb11fe3d4752a23a4e)

+++ src/main/java/de/maxhenkel/audioplayer/command/AudioPlayerCommands.java
@@ -559,11 +559,11 @@ public class AudioPlayerCommands {
     }
 
     private static FloatArgumentType maxMusicDiscRange() {
-        return FloatArgumentType.floatArg(1F, AudioPlayer.SERVER_CONFIG.maxMusicDiscRange.get().floatValue());
+        return FloatArgumentType.floatArg(1F, AudioPlayer.SERVER_CONFIG.maxMusicDiscRange.get());
     }
 
     private static FloatArgumentType maxGoatHornRange() {
-        return FloatArgumentType.floatArg(1F, AudioPlayer.SERVER_CONFIG.maxGoatHornRange.get().floatValue());
+        return FloatArgumentType.floatArg(1F, AudioPlayer.SERVER_CONFIG.maxGoatHornRange.get());
     }
 
 }
